requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudinary==1.29.0
colorama==0.4.4
coverage==6.3.3
crispy-bootstrap5==0.6
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-phone-field==1.8.1
gunicorn==20.1.0
idna==3.3
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
urllib3==1.26.9
